Removing programUtils and fixing small css issues
