+++ src/components/utils/icons.js
@@ -1,0 +1,18 @@
+import React from "react"
+import {
+  FaCalendarCheck,
+  FaCoffee,
+  FaRobot,
+  FaShieldAlt,
+  FaUserTie,
+} from "react-icons/fa"
+
+const icons = {
+  "calendar-check": <FaCalendarCheck color="#0c0044" size="2em" />,
+  coffee: <FaCoffee color="#0c0044" size="2.5em" />,
+  robot: <FaRobot />,
+  "shield-alt": <FaShieldAlt />,
+  "user-tie": <FaUserTie />,
+}
+
+export default icons
